refactor aggregation arbitration test
